Move the pgvector store in a dedicate spring-ai-pgvector-store module  - merge the spring-ai-it into spring-ai-pgvector-storee
