+++ spring-ai-core/pom.xml
@@ -20,28 +20,10 @@
 	</scm>
 
 	<properties>
-		<pgvector.version>0.1.3</pgvector.version>
-		<postgresql.version>42.6.0</postgresql.version>
 		<jsonschema.version>4.31.1</jsonschema.version>
 	</properties>
 
 	<dependencies>
-
-		<!-- PGVector -->
-		<dependency>
-			<groupId>com.pgvector</groupId>
-			<artifactId>pgvector</artifactId>
-			<version>${pgvector.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-jdbc</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.postgresql</groupId>
-			<artifactId>postgresql</artifactId>
-			<version>${postgresql.version}</version>
-		</dependency>
 
 		<!-- production dependencies -->
 		<dependency>
