+++ spring-ai-pgvector-store/src/test/java/org/springframework/ai/vectorstore/README.md
@@ -1,0 +1,16 @@
+
+# PGvector VectorStore
+
+Pgvector is an open-source extension for PostgreSQL that enables storing and searching over machine learning-generated embeddings. It provides different capabilities that let users identify both exact and approximate nearest neighbors. It is designed to work seamlessly with other PostgreSQL features, including indexing and querying.
+
+## Start Postgres+PGVecgor DB:
+
+```
+docker run -it --rm --name postgres -p 5432:5432 -e POSTGRES_USER=postgres -e POSTGRES_PASSWORD=postgres ankane/pgvector
+```
+
+You can connect to this server like this:
+
+```
+psql -U postgres -h localhost -p 5432
+```
